LayoutTests/fast/js/function-apply-many-args-expected.txt

Tests that we truncate arguments beyond a certain threshold.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f.apply(null) is 0
PASS f.apply(null, []) is 0
PASS f.apply(null, [1]) is 1
PASS f.apply(null, new Array(10)) is 10
PASS f.apply(null, new Array(1000)) is 1000
PASS f.apply(null, new Array(65536)) is 65536
PASS f.apply(null, new Array(65537)) is 65536
PASS f.apply(null, new Array(65537)) is 65536
PASS f.apply(null, bigArray) is 65536
PASS g.apply(null) is 0
PASS g.apply(null, []) is 0
PASS g.apply(null, [1]) is 1
PASS g.apply(null, new Array(10)) is 10
PASS g.apply(null, new Array(1000)) is 1000
PASS g.apply(null, new Array(65536)) is 65536
PASS g.apply(null, new Array(65537)) is 65536
PASS g.apply(null, new Array(65537)) is 65536
PASS g.apply(null, bigArray) is 65536
PASS h.apply(null) is 0
PASS h.apply(null, []) is 0
PASS h.apply(null, [1]) is 1
PASS h.apply(null, new Array(10)) is 10
PASS h.apply(null, new Array(1000)) is 1000
PASS h.apply(null, new Array(65536)) is 65536
PASS h.apply(null, new Array(65537)) is 65536
PASS h.apply(null, new Array(65537)) is 65536
PASS h.apply(null, bigArray) is 65536
PASS i.apply(null) is 0
PASS i.apply(null, []) is 0
PASS i.apply(null, [1]) is 1
PASS i.apply(null, new Array(10)) is 10
PASS i.apply(null, new Array(1000)) is 1000
PASS i.apply(null, new Array(65536)) is 65536
PASS i.apply(null, new Array(65537)) is 65536
PASS i.apply(null, new Array(65537)) is 65536
PASS i.apply(null, bigArray) is 65536
PASS successfullyParsed is true

TEST COMPLETE
